Add temporal Q-former training pipeline

+++ simlingo_training/models/language_model/llm.py
@@ -85,7 +85,11 @@ class LLM(nn.Module):
             self.model = self.model.language_model
             self.model.embed_tokens = self.model.base_model.embed_tokens
         elif 'internvl' in self.variant.lower():
-            self.model = AutoModel.from_pretrained(self.variant, trust_remote_code=True)
+            self.model = AutoModel.from_pretrained(
+                self.variant,
+                trust_remote_code=True,
+                torch_dtype=torch.float16,
+            )
             self.model = self.model.language_model
             try:
                 self.model.embed_tokens = self.model.base_model.embed_tokens
